+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-All notable changes to the `InterFAX notification channel` will be documented in this file
+All notable changes to the `InterFAX notification channel` will be documented in this file.
+
+## 2.1.1 - 2021-05-12
+
+- Prevent API limit requests with updates to status polling
+- Update exception handling for better error messages from InterFAX client
+
+## 2.1.0 - 2020-11-10
+
+- Add Laravel 8 support
 
 ## 2.0.0 - 2020-07-15